split tests, add missing tests
