+++ test/data/pins-car-sharing.json
@@ -1,0 +1,1 @@
+["poiID:2104108141:14612000:|c||teilAuto Station Nürnberger Straße - (NU)|5656855|4621006||","poiID:2104108096:14612000:|c||teilAuto Station Fritz-Löffler-Straße (Studentenwerk) - (FL)|5657104|4621471||","poiID:2104108140:14612000:|c||teilAuto Station Nöthnitzer Straße - (NO)|5656014|4620448||"]
