+++ test/data/pins-park-and-ride.json
@@ -1,0 +1,1 @@
+["poiID:2104107903:14628110:|pr|Freital|P+R Freital-Potschappel|5654410|4616752|10 Stellplätze, kostenfrei, 24h|","poiID:2104107857:14612000:|pr||P+R Dresden Pennrich|5657409|4614165|48 Stellplätze, kostenfrei, 24h|","poiID:2104107901:14628110:|pr|Freital|P+R Freital-Deuben|5652980|4615621|49 Stellplätze, kostenfrei, 24h|","poiID:2104107855:14612000:|pr||P+R Dresden Kaditz|5662033|4618583|195 Stellplätze, kostenfrei, 24h|","poiID:2104107902:14628110:|pr|Freital|P+R Freital-Hainsberg|5651662|4615191|110 Stellplätze, kostenfrei, 24h|"]
