+++ test/data/pins-rent-a-bike.json
@@ -1,0 +1,1 @@
+["poiID:2104108009:14612000:|r||SZ-Bike Station - Nürnberger Platz|5656570|4621200||","poiID:2104107979:14612000:|r||SZ-Bike Station - Nürnberger Ei|5656814|4620957||","poiID:2104107976:14612000:|r||SZ-Bike Station - Bergstr./Zellescher Weg|5656431|4621534||","poiID:2104107983:14612000:|r||SZ-Bike Station - Bautest|5656199|4620902||","poiID:2104107981:14612000:|r||SZ-Bike Station - Mommsenstr./Bergstr.|5656107|4621546||","poiID:2104108008:14612000:|r||SZ-Bike Station - Bamberger Straße|5656589|4620565||","poiID:2104107982:14612000:|r||SZ-Bike Station - Nöthnitzer Straße|5655898|4621211||","poiID:2104108007:14612000:|r||SZ-Bike Station - Leubnitzer Straße|5657252|4621112||","poiID:2104107975:14612000:|r||SZ-Bike Station - HTW Dresden/Hochschulstraße|5657020|4621775||","poiID:2104108010:14612000:|r||SZ-Bike Station - Plauen/Nöthnitzer Straße|5656124|4620487||"]
