+++ test/data/pins-ticket-machine.json
@@ -1,0 +1,1 @@
+["poiID:2104108200:14612000:|t||Ticketautomat Dresden, Nürnberger Platz -|5656639|4621187||","poiID:2104108302:14612000:|t||Ticketautomat Dresden, Münchner Platz -|5656386|4620914||","poiID:2104108255:14612000:|t||Ticketautomat Dresden, Südvorstadt -|5656841|4620905||","poiID:2104108239:14612000:|t||Ticketautomat Dresden, Reichenbachstraße -|5656927|4621532||","poiID:2104108257:14612000:|t||Ticketautomat Dresden, Nöthnitzer Straße -|5656165|4620532||","poiID:2104108241:14612000:|t||Ticketautomat Dresden, Chemnitzer Straße -|5657165|4620478||"]
