--- test/test.js
@@ -1,592 +1,0 @@
-'use strict';
-
-var fs = require('fs');
-var assert = require('assert');
-var mockery = require('mockery');
-var bluebird = require('bluebird');
-var dvb;
-
-function mockRequest(filename) {
-    before(function (done) {
-        if (process.env.NODE_ENV != 'test_live') {
-            mockery.enable({
-                warnOnReplace: true,
-                warnOnUnregistered: false,
-                useCleanCache: true
-            });
-            mockery.registerMock('request-promise', function () {
-                var response = fs.readFileSync(__dirname + '/data/' + filename, 'utf8');
-                return bluebird.resolve(response.trim());
-            });
-        }
-        dvb = require('../index');
-        done();
-    });
-
-    after(function (done) {
-        mockery.disable();
-        mockery.deregisterAll();
-        done();
-    });
-}
-
-describe('dvb.monitor', function () {
-    describe('dvb.monitor "Postplatz"', function () {
-        mockRequest('monitor-postplatz.json');
-
-        function assertTransport(transport) {
-            assert(transport.line);
-            assert(transport.direction);
-            assert.strictEqual('number', typeof transport.arrivalTimeRelative);
-            assert.strictEqual('object', typeof transport.arrivalTime);
-        }
-
-        it('should return an array with elements', function (done) {
-            dvb.monitor('postplatz', 0, 5)
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert(data.length > 0);
-                    done()
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should contain all four fields', function (done) {
-            dvb.monitor('postplatz', 0, 5)
-                .then(function (data) {
-                    data.forEach(assertTransport);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return a Promise but still accept a callback', function () {
-            dvb.monitor('postplatz', 0, 5, function (err, data) {
-                    assert(data);
-                })
-                .then(assert)
-        });
-    });
-
-    describe('dvb.monitor "xxx"', function () {
-        mockRequest('empty_json.json');
-
-        it('should return an empty array', function (done) {
-            dvb.monitor('xxx', 0, 5)
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert.equal(data.length, 0);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    })
-});
-
-describe('dvb.route', function () {
-    describe('dvb.route "Prager Straße -> Postplatz"', function () {
-        mockRequest('route-pragerstr-postplatz.json');
-
-        function assertTrip(trip) {
-            assert(trip.departure);
-            assert(trip.arrival);
-            assert(trip.duration);
-            assert.strictEqual('number', typeof trip.interchange);
-
-            assert(Array.isArray(trip.nodes));
-            trip.nodes.forEach(assertNode);
-        }
-
-        function assertNode(node) {
-            assert(node.mode);
-            assert(node.line || node.line === '');
-            assert(node.direction || node.direction === '');
-
-            assert.strictEqual('object', typeof node.departure);
-            assert(node.departure.stop);
-            assert(node.departure.time);
-
-            assert(Array.isArray(node.departure.coords));
-            assert.strictEqual(2, node.departure.coords.length);
-
-            assert.strictEqual('object', typeof node.arrival);
-            assert(node.arrival.stop);
-            assert(node.arrival.time);
-
-            assert(Array.isArray(node.arrival.coords));
-            assert.strictEqual(2, node.arrival.coords.length);
-
-            assert(Array.isArray(node.path));
-            node.path.forEach(function (path) {
-                assert(Array.isArray(path));
-                assert.strictEqual(2, path.length);
-                assert.strictEqual(51, Math.floor(path[0]));
-                assert.strictEqual(13, Math.floor(path[1]));
-            });
-        }
-
-        it('should return the correct origin and destination', function (done) {
-            dvb.route('pragerstrasse', 'postplatz', new Date(), dvb.route.DEPARTURE)
-                .then(function (data) {
-                    assert.strictEqual('Dresden, Prager Straße', data.origin);
-                    assert.strictEqual('Dresden, Postplatz', data.destination);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return an array of trips', function (done) {
-            dvb.route('pragerstrasse', 'postplatz', new Date(), dvb.route.DEPARTURE)
-                .then(function (data) {
-                    assert(Array.isArray(data.trips));
-                    assert(data.trips.length > 0);
-                    data.trips.forEach(assertTrip);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return a Promise but still accept a callback', function (done) {
-            dvb.route('pragerstrasse', 'postplatz', new Date(), dvb.route.DEPARTURE, function (err, data) {
-                assert(data);
-                done();
-            }).then(assert);
-        });
-    });
-
-    describe('dvb.route "0 -> 0"', function () {
-        mockRequest('empty_json.json');
-
-        it('should return null', function (done) {
-            dvb.route('0', '0', new Date(), dvb.route.DEPARTURE)
-                .then(function (data) {
-                    assert.equal(null, data);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-});
-
-describe('dvb.find', function () {
-    describe('dvb.find "Zellescher Weg"', function () {
-        mockRequest('find-zellescherweg.json');
-
-        function assertStop(stop) {
-            assert(stop.stop);
-            assert(Array.isArray(stop.coords));
-            assert.strictEqual(2, stop.coords.length);
-            assert.strictEqual(51, Math.floor(stop.coords[0]));
-            assert.strictEqual(13, Math.floor(stop.coords[1]));
-        }
-
-        it('should return an array', function (done) {
-            dvb.find('zellesch')
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert(data.length > 0);
-                    data.forEach(assertStop);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should find the correct stop', function (done) {
-            dvb.find('zellesch')
-                .then(function (data) {
-                    assert.strictEqual('Zellescher Weg', data[0].stop);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return a Promise but still accept a callback', function (done) {
-            dvb.find('zellesch', function (err, data) {
-                assert(data);
-                done();
-            }).then(assert);
-        });
-    });
-
-    describe('dvb.find "0"', function () {
-        mockRequest('empty_json.json');
-
-        it('should return an empty array', function (done) {
-            dvb.find('0')
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert.equal(data.length, 0);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    })
-});
-
-describe('dvb.pins', function () {
-    describe('dvb.pins "51.026578, 13.713899, 51.035565, 13.737974, stop"', function () {
-        mockRequest('pins-stop.json');
-
-        it('should resolve into an array', function (done) {
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.STOP, {})
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert.notEqual(0, data.length);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should contain objects with id, name, coords, type and connections', function (done) {
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.STOP, {})
-                .then(function (data) {
-                    data.forEach(function (elem) {
-                        assert(elem.id);
-                        assert(elem.name);
-                        assert(elem.coords);
-                        assert(elem.type);
-                        assert.strictEqual(2, elem.coords.length);
-                        assert.strictEqual(13, Math.floor(elem.coords[1]));
-                        assert.strictEqual(51, Math.floor(elem.coords[0]));
-                        assert.strictEqual(13, Math.floor(elem.coords[1]));
-                        assert(Array.isArray(elem.connections));
-                        elem.connections.forEach(function (con) {
-                            assert(con.line);
-                            assert(con.type);
-                        });
-                    });
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return a Promise but still accept a callback', function (done) {
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, 'stop', {}, function (err, data) {
-                assert(data);
-                done();
-            }).then(assert);
-        });
-
-        describe('test options', function () {
-
-            it('default: connections, not groued, only title of line type', function (done) {
-                var hasConnections = false;
-                dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.STOP, {})
-                    .then(function (data) {
-                        assert.notEqual(0, data.length);
-                        data.forEach(function (elem) {
-                            assert(elem.connections);
-                            if (elem.connections.length > 0) {
-                                hasConnections = true;
-                            }
-                            elem.connections.forEach(function (connection) {
-                                assert(connection.type);
-                                assert(connection.line);
-                                assert.equal(Array.isArray(connection.line), false);
-                            });
-                        });
-                        assert(hasConnections);
-                        done();
-                    })
-                    .catch(function (err) {
-                        done(err);
-                    });
-            });
-
-            it('showLines: false', function (done) {
-                dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.STOP, {showLines: false})
-                    .then(function (data) {
-                        assert.notEqual(0, data.length);
-                        data.forEach(function (elem) {
-                            assert.equal(elem.connections, undefined);
-                        });
-                        done();
-                    })
-                    .catch(function (err) {
-                        done(err);
-                    });
-            });
-
-            it('groupByType: true', function (done) {
-                dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.STOP, {groupByType: true})
-                    .then(function (data) {
-                        assert.notEqual(0, data.length);
-                        data.forEach(function (elem) {
-                            assert(elem.connections);
-                            elem.connections.forEach(function (connection) {
-                                assert(connection.type);
-                                assert(connection.lines);
-                                assert(Array.isArray(connection.lines));
-                            });
-                        });
-                        done();
-                    })
-                    .catch(function (err) {
-                        done(err);
-                    });
-            });
-
-            it('fullLineType: true', function (done) {
-                dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.STOP, {fullLineType: true})
-                    .then(function (data) {
-                        assert.notEqual(0, data.length);
-                        data.forEach(function (elem) {
-                            assert(elem.connections);
-                            elem.connections.forEach(function (connection) {
-                                assert(connection.type);
-                                assert(connection.type.name);
-                                assert(connection.type.title);
-                                assert(connection.type.icon_url);
-                            });
-                        });
-                        done();
-                    })
-                    .catch(function (err) {
-                        done(err);
-                    });
-            });
-
-        });
-    });
-
-    describe('dvb.pins "51.026578, 13.713899, 51.035565, 13.737974, platform"', function () {
-        mockRequest('pins-platform.json');
-
-        it('should contain objects with name, coords, type and platform_nr', function (done) {
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.PLATFORM, {})
-                .then(function (data) {
-                    assert.notEqual(0, data.length);
-                    data.forEach(function (elem) {
-                        assert(elem.name);
-                        assert(elem.coords);
-                        assert(elem.type);
-                        assert.strictEqual(2, elem.coords.length);
-                        assert.strictEqual(13, Math.floor(elem.coords[1]));
-                        assert.strictEqual(51, Math.floor(elem.coords[0]));
-                        assert(elem.platform_nr);
-                    });
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-
-    describe('dvb.pins "51.026578, 13.713899, 51.035565, 13.737974, POI"', function () {
-        mockRequest('pins-poi.json');
-
-        it('should contain objects with name, coords, type and id', function (done) {
-            var pinType = dvb.pins.type.POI;
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, pinType, {})
-                .then(function (data) {
-                    assert.notEqual(0, data.length);
-                    data.forEach(function (elem) {
-                        assert(elem.id);
-                        assert(elem.name);
-                        assert(elem.coords);
-                        assert(elem.type);
-                        assert.strictEqual(2, elem.coords.length);
-                        assert.strictEqual(13, Math.floor(elem.coords[1]));
-                        assert.strictEqual(51, Math.floor(elem.coords[0]));
-                    });
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-
-    describe('dvb.pins "51.026578, 13.713899, 51.035565, 13.737974, POI+PLATFORM"', function () {
-        mockRequest('pins-poi-pf.json');
-
-        it('should contain objects with pinType of RENT_A_BIKE or PLATFORM', function (done) {
-            var hasPOI = false, hasSTOP = false;
-
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, [dvb.pins.type.RENT_A_BIKE, dvb.pins.type.PLATFORM])
-                .then(function (data) {
-                    data.forEach(function (elem) {
-                        if (elem.type == dvb.pins.type.RENT_A_BIKE)
-                            hasPOI = true;
-                        else if (elem.type == dvb.pins.type.PLATFORM)
-                            hasSTOP = true;
-                        else
-                            assert.fail(elem.type, dvb.pins.type.RENT_A_BIKE + " || " + dvb.pins.type.PLATFORM, "pinType should be '" + dvb.pins.type.RENT_A_BIKE + " or '" + dvb.pins.type.PLATFORM + "'");
-                    });
-                    assert(hasPOI && hasSTOP, "pinTypes should be '" + dvb.pins.type.RENT_A_BIKE + " and '" + dvb.pins.type.PLATFORM + "'");
-
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-
-    describe('dvb.pins "0, 0, 0, 0, stop"', function () {
-        mockRequest('empty.json');
-
-        it('should resolve into an empty array', function (done) {
-            dvb.pins(0, 0, 0, 0, dvb.pins.type.STOP)
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert.equal(0, data.length);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-})
-;
-
-describe('dvb.address', function () {
-    describe('dvb.address "51.025451, 13.722943"', function () {
-        mockRequest('address-51-13.json');
-
-        it('should resolve into an object with city and address properties', function (done) {
-            dvb.address(51.025451, 13.722943)
-                .then(function (data) {
-                    assert.strictEqual("Nöthnitzer Straße 46", data.address);
-                    assert.strictEqual("Dresden", data.city);
-                    done()
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return a Promise but still accept a callback', function (done) {
-            dvb.address(51.025451, 13.722943, function (err, data) {
-                assert(data);
-                done();
-            }).then(assert);
-        });
-    });
-
-    describe('dvb.address "0, 0"', function () {
-        mockRequest('empty.json');
-
-        it('should return null', function (done) {
-            dvb.address(0, 0)
-                .then(function (data) {
-                    assert.equal(null, data);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-});
-
-describe('dvb.coords', function () {
-    describe('dvb.coords "33000755"', function () {
-        mockRequest('coords-33000755.json');
-
-        it('should resolve into a coordinate array [lat, lng]', function (done) {
-            dvb.coords('33000755')
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert.equal(data.length, 2);
-                    assert.strictEqual(13, Math.floor(data[1]));
-                    assert.strictEqual(51, Math.floor(data[0]));
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-
-        it('should return a Promise but still accept a callback', function (done) {
-            dvb.coords('33000755', function (err, data) {
-                assert(data);
-                done();
-            }).then(assert);
-        });
-    });
-
-    describe('dvb.coords "xxx"', function () {
-        mockRequest('empty.json');
-
-        it('should return null', function (done) {
-            dvb.coords("xxx")
-                .then(function (data) {
-                    assert.equal(null, data);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-});
-
-describe('dvb.coords for id from dvb.pins', function () {
-
-    var pins = [];
-
-    describe('dvb.pins "51.026578, 13.713899, 51.035565, 13.737974, POI"', function () {
-        mockRequest('pins-poi.json');
-
-        it('should contain objects with name, coords and id', function (done) {
-            dvb.pins(51.026578, 13.713899, 51.035565, 13.737974, dvb.pins.type.POI)
-                .then(function (data) {
-                    assert(Array.isArray(data));
-                    assert.notEqual(0, data.length);
-                    data.forEach(function (elem) {
-                        assert(elem.id);
-                        assert(elem.name);
-                        assert(elem.coords);
-                        assert.strictEqual(2, elem.coords.length);
-                        assert.strictEqual(13, Math.floor(elem.coords[1]));
-                        assert.strictEqual(51, Math.floor(elem.coords[0]));
-                    });
-                    pins = data;
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-
-    describe('dvb.coords should be equal to first pin coords', function () {
-        mockRequest('coords-poi.json');
-
-        it('coordinates should be equal', function (done) {
-            dvb.coords(pins[0].id)
-                .then(function (coords) {
-                    assert.deepEqual(coords, pins[0].coords);
-                    done();
-                })
-                .catch(function (err) {
-                    done(err);
-                });
-        });
-    });
-});
